airdrop-farm/requirements.txt
requests==2.31.0
colorama==0.4.6
rich==13.7.0
pyfiglet==1.0.2
tqdm==4.66.1
pysocks==1.7.1
fake-useragent==1.4.0
asyncio-throttle==1.0.2
aiohttp==3.9.1
aiofiles==23.2.0
aiohttp-socks==0.8.4